+++ lib/internal/http.rs
@@ -115,6 +115,12 @@ pub fn get_request_headers(raw_data: &str) -> Vec<Header> {
     headers
 }
 
+// TODO: Test This
+/// Check if the socket connetion wants to use keep alive
+pub fn connection_mode(headers: &[Header]) -> bool {
+    matches!(headers.iter().find(|x| x.name == "Connection"), Some(i) if i.value == "keep-alive")
+}
+
 /// Get Cookies of a raw HTTP request.
 #[cfg(feature = "cookies")]
 pub fn get_request_cookies(raw_data: &str) -> Vec<Cookie> {