+++ Changelog.md
@@ -1,5 +1,7 @@
 # 1.2.0
 
+_comming soon?_
+
 - oh windows,,,
 - Fix Path Traversal on windows
 - Use AsRef<str> more instead of Display
@@ -17,6 +19,8 @@
 
 # 1.1.0
 
+Apr 10, 2022
+
 - Update Path Matcher to support AnyAfter segments (\*\*)
 - Remove Test Example
 - Add Paste Bin App Example
@@ -24,6 +28,8 @@
 - Fix Path Traversal Exploit O_O
 
 # 1.0.0!
+
+Mar 14, 2022
 
 - Add ThreadPool Back!
 - Tracing Feature
@@ -33,6 +39,8 @@
 - Make use of end middleware on logger
 
 # 0.4.0
+
+Feb 19, 2022
 
 - Make serve static types public
 - Fix File Uploading stuff
@@ -48,6 +56,8 @@
 - Now .header takes in a key and value rather than a Header struct
 
 # 0.3.0
+
+Jan 25, 2022 s
 
 - Add content types
 - Update Logger syntax
@@ -77,11 +87,15 @@
 
 # 0.2.2
 
+Dec 04, 2021
+
 - Remove Debug Print Left in...
 - Small changes to features
 - This is mostly about the Debug Print
 
 # 0.2.1
+
+Dec 04, 2021
 
 - Only Build common::remove_address_port if logger or rate-limiter are enabled
 - Make Header name / value Public
@@ -103,6 +117,8 @@
 
 # 0.2.0
 
+Nov 04, 2021
+
 - Response Overhaul, Now more like a Response Builder
 - Update _every_ example with new syntax...
 - Small improvement to Query parsing
@@ -118,6 +134,8 @@
 
 # 0.1.7
 
+Oct 27, 2021
+
 - Add Panic Message to Error Handel
 - Add http.rs to move raw http parsing out of server.rs
 - Start / Start Threaded returns Option
@@ -128,6 +146,8 @@
 - Add Socket Timeout Docs
 
 # 0.1.6
+
+Oct 20, 2021
 
 - Add Example for Logging
 - Add Example for Rate Limiter
@@ -147,6 +167,8 @@
 
 # 0.1.5
 
+Sep 17, 2021
+
 - Add a route error handler
 - Add `set_error_handler` fn to set the error handler
 - Implement clone for more structs
@@ -163,19 +185,27 @@
 
 # 0.1.4
 
+Sep 05, 2021
+
 - Allow responding with bytes, not just strings
 - Add Serving Favoricon as example thing
 
 # 0.1.3
+
+Sep 03, 2021
 
 - Add Support for query strings in paths
 - Add More docs for Query
 
 # 0.1.2
 
+Sep 01, 2021
+
 - Fix a bug where '.any' routes were not working
 
 # 0.1.1
+
+Aug 31, 2021
 
 - Add Optional Rate limiter _beta_
 - Update readme abit
@@ -189,5 +219,7 @@
 - Code Cleanup
 
 # 0.1.0
+
+Aug 21, 2021
 
 - Base Code