+++ Changelog.md
@@ -1,4 +1,6 @@
-# 1.1.1
+# 1.2.0
+
+_coming soon?_
 
 - oh windows,,,
 - Fix Path Traversal on windows
@@ -11,8 +13,15 @@
 - Improve built-in serve_static middleware
 - Re organize extension stuff
 - RateLimit use RwLock
+- Add Request ID Middleware
+- Server Wide State
+- Add Cache Middleware
+- Remove insane build script
+- When building http response only add Content-Length and default headers if they are not already present
 
 # 1.1.0
+
+Apr 10, 2022
 
 - Update Path Matcher to support AnyAfter segments (\*\*)
 - Remove Test Example
@@ -22,6 +31,8 @@
 
 # 1.0.0!
 
+Mar 14, 2022
+
 - Add ThreadPool Back!
 - Tracing Feature
 - Remove Middleware Interior Mutability by Default
@@ -30,6 +41,8 @@
 - Make use of end middleware on logger
 
 # 0.4.0
+
+Feb 19, 2022
 
 - Make serve static types public
 - Fix File Uploading stuff
@@ -45,6 +58,8 @@
 - Now .header takes in a key and value rather than a Header struct
 
 # 0.3.0
+
+Jan 25, 2022 s
 
 - Add content types
 - Update Logger syntax
@@ -74,11 +89,15 @@
 
 # 0.2.2
 
+Dec 04, 2021
+
 - Remove Debug Print Left in...
 - Small changes to features
 - This is mostly about the Debug Print
 
 # 0.2.1
+
+Dec 04, 2021
 
 - Only Build common::remove_address_port if logger or rate-limiter are enabled
 - Make Header name / value Public
@@ -100,6 +119,8 @@
 
 # 0.2.0
 
+Nov 04, 2021
+
 - Response Overhaul, Now more like a Response Builder
 - Update _every_ example with new syntax...
 - Small improvement to Query parsing
@@ -115,6 +136,8 @@
 
 # 0.1.7
 
+Oct 27, 2021
+
 - Add Panic Message to Error Handel
 - Add http.rs to move raw http parsing out of server.rs
 - Start / Start Threaded returns Option
@@ -125,6 +148,8 @@
 - Add Socket Timeout Docs
 
 # 0.1.6
+
+Oct 20, 2021
 
 - Add Example for Logging
 - Add Example for Rate Limiter
@@ -144,6 +169,8 @@
 
 # 0.1.5
 
+Sep 17, 2021
+
 - Add a route error handler
 - Add `set_error_handler` fn to set the error handler
 - Implement clone for more structs
@@ -160,19 +187,27 @@
 
 # 0.1.4
 
+Sep 05, 2021
+
 - Allow responding with bytes, not just strings
 - Add Serving Favoricon as example thing
 
 # 0.1.3
+
+Sep 03, 2021
 
 - Add Support for query strings in paths
 - Add More docs for Query
 
 # 0.1.2
 
+Sep 01, 2021
+
 - Fix a bug where '.any' routes were not working
 
 # 0.1.1
+
+Aug 31, 2021
 
 - Add Optional Rate limiter _beta_
 - Update readme abit
@@ -186,5 +221,7 @@
 - Code Cleanup
 
 # 0.1.0
+
+Aug 21, 2021
 
 - Base Code